put popup on order details
