+++ packages/p2p/src/components/orders/order-details/order-details.scss
@@ -11,7 +11,7 @@
             color: var(--text-general);
 
             &--success {
-                color: var(--text-profit-success)
+                color: var(--text-profit-success);
             }
         }
         &--outer {
