+++ packages/p2p/src/components/orders/orders.jsx
@@ -1,25 +1,16 @@
 import React        from 'react';
-import { Dialog } from 'deriv-components';
 import { localize } from 'deriv-translations';
 import PageReturn   from 'Components/page-return/page-return.jsx';
 import OrderDetails from './order-details/order-details.jsx';
 import OrderInfo    from './order-info';
-import Popup        from './popup.jsx';
 import './orders.scss';
 
 const Orders = () => {
     const order_info = new OrderInfo();
     const [order_details, setDetails] = React.useState(order_info);
-    const [show_popup, setShowPopup] = React.useState(false);
-    const [popup_options, setPopupOptions] = React.useState({});
+
     // TODO: [p2p-replace-with-api] - remove these dev toggle once data fetch works
     const toggleDetails = () => order_details && order_details.order_id ? setDetails(null) : setDetails(order_info);
-    const onCancelClick = () => setShowPopup(false);
-
-    const handleShowPopup = (options) => {
-        setPopupOptions(options);
-        setShowPopup(true);
-    }
 
     return (
         <div className='orders'>
@@ -36,25 +27,9 @@ const Orders = () => {
                     />
                     <OrderDetails
                         order_details={ order_details }
-                        showPopup={handleShowPopup}
                     />
                 </React.Fragment>
             }
-            {show_popup && (
-                <div className='orders__dialog'>
-                    <Dialog
-                        is_visible={show_popup}
-                        disableApp={() => {
-                            /* do nothing // disableApp is a mandatory props in dialog */
-                        }}
-                        enableApp={() => {
-                            /* do nothing // enableApp is a mandatory props in dialog */
-                        }}
-                    >
-                        <Popup {...popup_options} onCancel={onCancelClick} />
-                    </Dialog>
-                </div>
-            )}
         </div>
     );
 };
