+++ packages/p2p/src/components/orders/orders.scss
@@ -1,6 +1,5 @@
 .orders {
     padding: 2.4rem;
-    display: flex;
 
     &__button {
         margin: 0 1.6rem;
